feat #56: 지도 컴포넌트의 훅 분리/ 지도 관련 컴포넌트 리네이밍
