+++ src/constants/map/mapConfig.ts
@@ -1,0 +1,23 @@
+export const MAP_DEFAULT_CENTER = { lat: 37.550701948532236, lng: 127.07428227734258 };
+export const MAP_LEVEL = 2;
+export const MAP_MIN_LEVEL = 1;
+export const MAP_MAX_LEVEL = 2;
+
+export const ALLOWED_BOUNDARY = {
+  sw: { lat: 37.547, lng: 127.0655 },
+  ne: { lat: 37.5545, lng: 127.084 },
+};
+
+export const OUTER_WORLD = [
+  { lat: 85, lng: -179.999 },
+  { lat: 85, lng: 179.999 },
+  { lat: -85, lng: 179.999 },
+  { lat: -85, lng: -179.999 },
+];
+
+export const SEJONG_UNIVERSITY_CAMPUS = [
+  { lat: 37.545, lng: 127.07 }, // SW
+  { lat: 37.5575, lng: 127.07 }, // NW
+  { lat: 37.5575, lng: 127.078 }, // NE
+  { lat: 37.545, lng: 127.078 }, // SE
+];
